casi listo

+++ src/Dashboard.jsx
@@ -52,25 +52,48 @@ const Dashboard=()=> {
     .then(() => {
       window.alert("Usuario eliminado con éxito");
       dispatch(getUsers());
-      dispatch(getAposts());
-      dispatch(getService());
-      dispatch(getContract());
-      dispatch(getDeletedAdpost());
       dispatch(getDeletedUsers());
-      dispatch(getDeletedContract());
     })
     .catch((error) => {
       console.log(error);
     });
   };
   
-  //DELETE USERS -------------------------------
+  const deleteAdpost = async (id)=>{
+    axios.delete(`https://promanitasapi.onrender.com/api/v1/adposts/${id}`)
+    .then(() => {
+      window.alert("Posteo eliminado con éxito");
+      dispatch(getAposts());
+      dispatch(getDeletedAdpost());
+    })
+    .catch((error) => {
+      console.log(error);
+    });
+  };
 
-  const confirmDelete = (id) => {
+  const deleteContract = async (id)=>{
+    axios.delete(`https://promanitasapi.onrender.com/api/v1/contract/${id}`)
+    .then(() => {
+      window.alert("Contrato eliminado con éxito");
+      dispatch(getContract());
+      dispatch(getDeletedContract());
+    })
+    .catch((error) => {
+      console.log(error);
+    });
+  };
+
+  const confirmDelete = (id,selector) => {
     const confirmMessage = "¿Realmente deseas eliminar?";
     const confirmed = window.confirm(confirmMessage);
-    if (confirmed) {
+    if (confirmed && selector === 'user') {
       deleteUser(id);
+    }
+    if (confirmed && selector === 'adpost') {
+      deleteAdpost(id);
+    }
+    if (confirmed && selector === 'contract') {
+      deleteContract(id);
     }
   };
 
@@ -83,11 +106,11 @@ const Dashboard=()=> {
       <button className="dashboard-btn" onClick={ e => setStates(e, allContracts)} >CONTRACTS</button>
        
       <div className="divbox">
-        <h1 className="h1box">Usuarios: {allUsers.length}</h1>
+        <h1 className="h1box">Usuarios activos: {allUsers.length}</h1>
         <h1 className="h1box">Usuarios eliminados: {allDeletedUsers.length}</h1>
-        <h1 className="h1box">Posteos: {allAdposts.length}</h1>
+        <h1 className="h1box">Posteos activos: {allAdposts.length}</h1>
         <h1 className="h1box">Posteos eliminados: {allDeletedAdposts.length}</h1>
-        <h1 className="h1box">Contratos:{allContracts.length}</h1>
+        <h1 className="h1box">Contratos activos: {allContracts.length}</h1>
         <h1 className="h1box">Contratos eliminados: {allDeletedContracts.length}</h1>
       </div>
 
@@ -101,7 +124,7 @@ const Dashboard=()=> {
                 <th>ID</th>
                 <th>Name</th>
                 <th>Info</th>
-                <th>Eliminar</th>
+                {infServer[0]?.description? <th>Eliminar</th> : null }
               </>
             ) : infServer[0]?.username ?(
               <>
@@ -121,6 +144,7 @@ const Dashboard=()=> {
                 <th>Fecha de inicio</th>
                 <th>Fecha de cierre</th>
                 <th>Precio</th>
+                <th>Detalles</th>
                 <th>Eliminar</th>
               </>
             )}
@@ -134,7 +158,9 @@ const Dashboard=()=> {
                 <td>{item.name}</td>
                 <td>{item.description || item.image}</td>
                 <td>
-                  <button onClick={() => confirmDelete(item.id)}>❎</button>
+                  { !item.description ? null
+                    :  <button onClick={() => confirmDelete(item.id,'adpost')}>❎</button>
+                    }
                 </td>
               </tr>
             ) 
@@ -149,7 +175,7 @@ const Dashboard=()=> {
                 <td>{item.address}</td>
                 <td>{item.password}</td>
                 <td>
-                  <button onClick={() => confirmDelete(item.id)}>❎</button>
+                  <button onClick={() => confirmDelete(item.id,'user')}>❎</button>
                 </td>
               </tr>
             ) : (
@@ -157,9 +183,10 @@ const Dashboard=()=> {
                 <td>{item.id}</td>
                 <td>{item.commencementDate.slice(0, 10)}</td>
                 <td>{item.terminationDate.slice(0, 10)}</td>
-                <td>{item.payment}</td>
+                <td>{item.payment} USD</td>
+                <td>{item.detail}</td>
                 <td>
-                  <button onClick={() => confirmDelete(item.id)}>❎</button>
+                  <button onClick={() => confirmDelete(item.id,'contract')}>❎</button>
                 </td>
               </tr>
             )
